+++ collector/src/main/java/com/bitmovin/analytics/data/LegacyErrorData.kt
@@ -4,4 +4,4 @@ import androidx.annotation.Keep
 
 // LegacyErrorData is stored unindexed as error_data in crate
 @Keep // Protect from obfuscation in case customers are using proguard
-data class LegacyErrorData(val msg: String, val details: Array<String> = emptyArray())
+data class LegacyErrorData(val msg: String, val details: List<String> = emptyList())